Minor fixes.
