+++ hal/network/lwip/esp32/esp32ncpnetif.h
@@ -42,6 +42,8 @@ public:
     virtual int powerUp() override;
     virtual int powerDown() override;
 
+    virtual int getPowerState(if_power_state_t* state) const override;
+
     static int ncpDataHandlerCb(int id, const uint8_t* data, size_t size, void* ctx);
     static void ncpEventHandlerCb(const NcpEvent& ev, void* ctx);
